pyproject.toml
[project]
name = "tds-project1-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "google>=3.0.0",
    "google-genai>=1.20.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "numpy>=2.2.6",
    "openai>=1.86.0",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

